requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.4.post1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
folium==0.14.0
geographiclib==2.0
geopandas==0.14.0
geopy==2.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
networkx==3.1
numpy==1.26.0
osmnx==1.6.0
packaging==23.1
pandas==2.1.1
pip==23.2.1
psycopg2-binary==2.9.7
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
sendgrid==6.10.0
setuptools==68.1.2
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.20
starkbank-ecdsa==2.2.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
